pyproject.toml
[tool.poetry]
name = "itsnp bot series"
version = "0.1.0"
description = ""
authors = ["Nishant Sapkota <snishant306@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
hikari = "^2.0.0-alpha.103"
hikari-tanjun = "^2.1.2-alpha.1"
tortoise-orm = "0.17.4"
aerich = "0.5.3"
asyncpg = "^0.24.0"
python-dotenv = "^0.19.1"

[tool.poetry.dev-dependencies]
black = {version = "^21.9b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
